Conflicts:
	plugins/hypervisors/ovm3/src/com/cloud/hypervisor/ovm3/hypervisor/Ovm3ResourceBase.java

+++ engine/orchestration/src/com/cloud/agent/manager/AgentAttache.java
@@ -126,7 +126,7 @@ public abstract class AgentAttache {
         _maintenance = maintenance;
         _requests = new LinkedList<Request>();
         _agentMgr = agentMgr;
-        _nextSequence = new Long(s_rand.nextInt(Short.MAX_VALUE)) << 48;
+        _nextSequence = new Long(s_rand.nextInt(Short.MAX_VALUE)).longValue() << 48;
     }
 
     public synchronized long getNextSequence() {
